requirements.txt
asgiref==3.2.10
Django==3.1.2
django-redis==4.12.1
django-redis-cache==3.0.0
djangorestframework==3.14.0
pytz==2020.1
redis==3.5.3
sqlparse==0.4.1
